agents/requirements.txt
crewai
groq
pypdf2
python-dotenv
fastapi
uvicorn